requirements.txt
Flask==3.0.0
googletrans==4.0.0-rc1
gTTS==2.4.0
pytesseract
PyPDF2>=3.0.0
python-docx
sumy
urllib3
nltk
gunicorn
pymupdf
opencv-python
fpdf
pydub
pyspellchecker
pyenchant==3.2.2
language_tool_python
spacy
scikit-learn
langid
speechrecognition
Flask-CORS
autocorrect
Flask-SQLAlchemy
bcrypt
Flask-Cors
flask_mail
